move operations tests to e2e suite
